# perf: improve non-SIMD with wordwise validation 