@seanmonstar 2022-09-08
This is awesome! Do you know how new of a compiler it requires? The MSRV CI job is grumpy 🤷 .