@paolobarbolini 2022-09-16
Looks like 1.44 https://doc.rust-lang.org/nightly/std/primitive.i16.html#method.from_ne_bytes