@AaronO 2023-04-13
@seanmonstar The SIMD-related PR is up here: https://github.com/seanmonstar/httparse/pull/131